📝 Refactor ModifierKeys to treat L/R separately
To avoid all the extra code repetition this would induce, this PR changes the underlying implementation to use a flags enum. Note that this is hidden as an implementation detail, so we can keep providing the rich layer of functionality on top.

Another change that had to be made is that the `IsSupersetOf` is now ambiguous. In most cases, you just want to check if - say - a shift button is pressed, not caring about which side is really pressed. This is why I kept both this logic, and added a strict superset check too.