✨ Make ModifierKeys usable outside Bearded.UI library (#16)
* :sparkles: Make it possible to instantiate empty ModifierKeys instance

* :sparkles: Create ModifierKeys builder

* :sparkles: Implement equality methods for ModifierKeys

* :pencil: Change ModifierKeys to a readonly struct

* :bug: Fix equality checks for ModifierKeys

* :bug: Set language version to C# 8.0

* :bug: Update Visual Studio version in appveyor config

* 📝 Refactor ModifierKeys to treat L/R separately

To avoid all the extra code repetition this would induce, this PR changes the underlying implementation to use a flags enum. Note that this is hidden as an implementation detail, so we can keep providing the rich layer of functionality on top.

Another change that had to be made is that the `IsSupersetOf` is now ambiguous. In most cases, you just want to check if - say - a shift button is pressed, not caring about which side is really pressed. This is why I kept both this logic, and added a strict superset check too.

* Revert "📝 Refactor ModifierKeys to treat L/R separately"

This reverts commit d61d0b47b5f80957219a172c9a1c87bd69ae22ea.

* ⛏ Replace builder with modification methods

* Update Bearded.UI/EventArgs/ModifierKeys.cs

Co-authored-by: Paul Scharf <amul@amulware.net>

Co-authored-by: Paul Scharf <amul@amulware.net>